enhancement: add another field
