+++ packages/shared/lib/contexts/onboarding/interfaces/onboarding-profile.interface.ts
@@ -13,5 +13,5 @@ export interface IOnboardingProfile extends IPersistedProfile {
     strongholdPassword?: string
     mustVisitProfileName?: boolean
     hasStoredMnemonic?: boolean
-    hasRecoveredProfile?: boolean
+    hasInitialisedProfileManager?: boolean
 }
